Put stuff in the right place
