+++ lib/style.js
@@ -15,6 +15,7 @@ var zlib = require('zlib');
 var tilelive = require('tilelive');
 var url = require('url');
 var source = require('./source');
+var upload = require('mapbox-upload');
 
 var defaults = {
     name:'',
@@ -374,3 +375,44 @@ style.error = function(id, err) {
     return cache[id].errors;
 };
 
+style.upload = function(id, callback) {
+    if (style.tmpid(id))
+        return callback(new Error('Style must be saved first'));
+    if (typeof tm.db._docs.user.plan.tm2z == 'undefined' || !tm.db._docs.user.plan.tm2z)
+        return callback(new Error('You are not allowed access to tm2z uploads yet.'));
+
+    style.info(id, function(err, data) {
+        if (err) return callback(err);
+        var mapid = (typeof data._prefs.mapid !== 'undefined' && data._prefs.mapid !== '')
+            ? data._prefs.mapid
+            : tm.db._docs.oauth.account + '.' + crypto.createHash('md5')
+                .update(new Date().getTime().toString())
+                .digest('hex')
+                .substr(0,8);
+
+        var pckage = tm._config.cache + '/package-' + mapid + '.tm2z';
+        style.toPackage(id, pckage, function(err) {
+            if (err) return callback(err);
+            upload({
+                file: pckage,
+                account: tm.db._docs.oauth.account,
+                accesstoken: tm.db._docs.oauth.accesstoken,
+                mapid: mapid
+            }, function(err, task) {
+                task.once('error', function(err) {
+                    return callback(err);
+                });
+                task.once('end', mapSaved);
+            });
+        });
+
+        function mapSaved() {
+            data._prefs.mapid = mapid;
+            style.save(data, function(){
+                fs.unlink(pckage, function() {
+                    return callback();
+                });
+            });
+        }
+    });
+};
